seperate util class for messages clean up with tool calls
