+++ llm_agent/__init__.py
@@ -1,2 +1,3 @@
 from . import models
 from . import controllers
+from . import utils
